Fix board cards dnd (#1263)
* refactor

* fix dnd and cards refactor

* remove moveCard disabling

* avatar glitch fix

* fix image hook

* remove test

* styleguide refactor

* styleguide fixes

* Task users list in memo, sidebar memo, sidebar header image memo

* custom memo for SafeImage
