+++ wondrous-app/components/Common/BoardColumn/index.ts
@@ -1,0 +1,3 @@
+import BoardColumn from './BoardColumn';
+
+export default BoardColumn;
