+++ wondrous-app/components/Common/Image/index.ts
@@ -1,0 +1,4 @@
+import SafeImage from './SafeImage';
+import DefaultUserImage from './DefaultUserImage';
+
+export { SafeImage, DefaultUserImage };
