+++ wondrous-app/components/Common/SidebarEntity/index.tsx
@@ -1,11 +1,14 @@
+import { useMemo } from 'react';
+import { useRouter } from 'next/router';
+
+import useSideBar from 'hooks/useSideBar';
+
 import CollapseExpandButton from 'components/Common/SidebarCollapseButton';
 import AboutEntity from 'components/Common/SidebarEntityAbout';
 import List from 'components/Common/SidebarEntityList';
 import RolesSidebar from 'components/Common/SidebarEntityRoles';
 import CollabsSidebar from 'components/Common/SidebarEntityCollabs';
 import { ChildrenWrapper, SidebarContent, SidebarWrapper, Wrapper } from 'components/Common/SidebarStyles';
-import useSideBar from 'hooks/useSideBar';
-import { useRouter } from 'next/router';
 
 const SIDEBAR_COMPONENTS = {
   collabs: () => <CollabsSidebar />,
@@ -16,15 +19,21 @@ const EntitySidebar = ({ children }) => {
   const { minimized } = useSideBar();
   const { query } = useRouter();
 
-  let Sidebar = () => (
-    <>
-      <AboutEntity />
-      <List />
-    </>
-  );
+  const Sidebar = useMemo(() => {
+    if (query.roles) {
+      return SIDEBAR_COMPONENTS.roles;
+    }
+    if (query.collabs) {
+      return SIDEBAR_COMPONENTS.collabs;
+    }
 
-  if (query.roles) Sidebar = SIDEBAR_COMPONENTS.roles;
-  if (query.collabs) Sidebar = SIDEBAR_COMPONENTS.collabs;
+    return () => (
+      <>
+        <AboutEntity />
+        <List />
+      </>
+    );
+  }, [query.roles, query.collabs]);
 
   return (
     <Wrapper>
