+++ wondrous-app/components/Common/Task/Card.tsx
@@ -1,0 +1,7 @@
+import ProposalCard from './ProposalCard';
+import TaskCard from './TaskCard';
+
+export default function Card(props) {
+  const { task } = props;
+  return task.isProposal ? <ProposalCard {...props} /> : <TaskCard {...props} />;
+}
