+++ wondrous-app/components/Common/Task/Task.tsx
@@ -1,47 +1,25 @@
-import React, { useCallback, useContext, useEffect, useState } from 'react';
+import React, { useCallback, useContext, useEffect, useState, memo, FC, useMemo } from 'react';
 import { useLazyQuery, useMutation } from '@apollo/client';
 import { useRouter } from 'next/router';
-import { cloneDeep } from 'lodash';
-import { useMe } from 'components/Auth/withAuth';
-import { ArchiveTaskModal } from 'components/Common/ArchiveTaskModal';
-import { AvatarList } from 'components/Common/AvatarList';
-import Compensation from 'components/Common/Compensation';
-import DeleteTaskModal from 'components/Common/DeleteTaskModal';
-import TaskViewModal from 'components/Common/TaskViewModal';
-import { SafeImage } from 'components/Common/Image';
-import { TaskMedia } from 'components/Common/MediaPlayer';
-import { SnackbarAlertContext } from 'components/Common/SnackbarAlert';
-import { flexDivStyle, rejectIconStyle } from 'components/Common/TaskSummary';
-import { TaskSummaryAction } from 'components/Common/TaskSummary/styles';
-import { Arrow } from 'components/Icons/sections';
-import { CompletedIcon } from 'components/Icons/statusIcons';
-import { RejectIcon } from 'components/Icons/taskModalIcons';
-import { CreateEntity } from 'components/CreateEntity';
-import { renderMentionString } from 'utils/common';
+import cloneDeep from 'lodash/cloneDeep';
+
+import { delQuery } from 'utils';
 import * as Constants from 'utils/constants';
 import { parseUserPermissionContext } from 'utils/helpers';
 import { useOrgBoard, usePodBoard, useUserBoard } from 'utils/hooks';
-import { delQuery } from 'utils';
 import { UPDATE_TASK_ASSIGNEE, ARCHIVE_TASK, UNARCHIVE_TASK } from 'graphql/mutations/task';
 import { GET_TASK_BY_ID, GET_TASK_REVIEWERS } from 'graphql/queries';
 import { CLOSE_TASK_PROPOSAL } from 'graphql/mutations/taskProposal';
+import type { TaskInterface } from 'types/task';
 
-import { useLocation } from 'utils/useLocation';
-import Card from './card';
-import TASK_ICONS from './constants';
-import {
-  ArchivedTaskUndo,
-  PodName,
-  PodWrapper,
-  TaskCardDescriptionText,
-  TaskContent,
-  TaskFooter,
-  TaskHeader,
-  TaskListCardWrapper,
-  TaskSeparator,
-  TaskStatusHeaderText,
-  TaskTitle,
-} from './styles';
+import { useMe } from 'components/Auth/withAuth';
+import { ArchiveTaskModal } from 'components/Common/ArchiveTaskModal';
+import DeleteTaskModal from 'components/Common/DeleteTaskModal';
+import { SnackbarAlertContext } from 'components/Common/SnackbarAlert';
+import { CreateEntity } from 'components/CreateEntity';
+
+import Card from './Card';
+import { ArchivedTaskUndo } from './styles';
 
 const getBoardType = ({ orgBoard, podBoard, userBoard }) => {
   if (orgBoard) return Constants.BOARD_TYPE.org;
@@ -78,7 +56,12 @@ const useGetTaskById = (editTask, task) => {
   return taskData?.getTaskById;
 };
 
-export function Task(props) {
+export type TaskProps = {
+  className?: string;
+  task: TaskInterface;
+};
+
+const Task: FC<TaskProps> = (props) => {
   // WHAT IS THIS COMPONENT? IS IT TASK CARD? POORLY NAMED
   const { task, className } = props;
   const {
@@ -91,7 +74,6 @@ export function Task(props) {
     assigneeId = null,
     assigneeUsername = null,
     assigneeProfilePicture = null,
-    users = [],
     type,
     createdBy,
     commentCount,
@@ -108,7 +90,6 @@ export function Task(props) {
   const user = useMe();
   const userPermissionsContext =
     orgBoard?.userPermissionsContext || podBoard?.userPermissionsContext || userBoard?.userPermissionsContext;
-  const [userList, setUserList] = useState([]);
   const [archiveTask, setArchiveTask] = useState(false);
   const [deleteTask, setDeleteTask] = useState(false);
   const [initialStatus, setInitialStatus] = useState('');
@@ -246,27 +227,24 @@ export function Task(props) {
     });
   };
 
-  useEffect(() => {
-    // One assigned person.
-    if (assigneeUsername) {
-      // clean
-      setUserList([
-        {
-          id: assigneeId,
-          name: assigneeUsername,
-          initials: assigneeUsername[0].toUpperCase(),
-          avatar: {
-            url: assigneeProfilePicture,
-            isOwnerOfPod: false,
-            color: null,
-          },
-        },
-      ]);
-    } else {
-      setUserList(users);
+  const userList = useMemo(() => {
+    if (!assigneeUsername) {
+      return [];
     }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [assigneeUsername]);
+
+    return [
+      {
+        id: assigneeId,
+        name: assigneeUsername,
+        initials: assigneeUsername[0].toUpperCase(),
+        avatar: {
+          url: assigneeProfilePicture,
+          isOwnerOfPod: false,
+          color: null,
+        },
+      },
+    ];
+  }, [assigneeUsername, assigneeId, assigneeProfilePicture]);
 
   return (
     <span className={className}>
@@ -342,119 +320,6 @@ export function Task(props) {
       />
     </span>
   );
-}
+};
 
-export function TaskListCard(props) {
-  const { taskType, task } = props;
-  const location = useLocation();
-  const router = useRouter();
-  const [viewDetails, setViewDetails] = useState(false);
-  const TaskIcon = () => {
-    if (task?.paymentStatus === Constants.PAYMENT_STATUS.PROCESSING) {
-      return TASK_ICONS[Constants.TASK_STATUS_AWAITING_PAYMENT];
-    }
-    if (task?.paymentStatus === Constants.PAYMENT_STATUS.PAID) {
-      return TASK_ICONS[Constants.TASK_STATUS_PAID];
-    }
-    return TASK_ICONS?.[taskType];
-  };
-
-  return (
-    <>
-      <TaskViewModal
-        open={viewDetails}
-        handleClose={() => {
-          setViewDetails(false);
-          const newUrl = `${delQuery(router.asPath)}?view=${router?.query?.view || 'grid'}&entity=${
-            location?.params?.entity || Constants.ENTITIES_TYPES.TASK
-          }`;
-          location.push(newUrl);
-        }}
-        taskId={taskType === Constants.TASK_STATUS_IN_REVIEW ? task?.taskId : task?.id}
-        isTaskProposal={taskType === Constants.TASK_STATUS_REQUESTED}
-        back
-      />
-      <TaskListCardWrapper
-        onClick={() => {
-          setViewDetails(true);
-        }}
-      >
-        <TaskHeader>
-          <SafeImage
-            useNextImage={false}
-            src={task?.orgProfilePicture}
-            style={{
-              width: '29px',
-              height: '28px',
-              borderRadius: '4px',
-            }}
-            alt="Organization logo"
-          />
-          <AvatarList
-            style={{ marginLeft: '12px' }}
-            users={[
-              {
-                id: task?.assigneeId || task?.createdBy,
-                name: task?.assigneeUsername || task?.creatorUsername,
-                initials:
-                  (task?.assigneeUsername && task?.assigneeUsername[0].toUpperCase()) ||
-                  (task?.creatorUsername && task?.creatorUsername[0].toUpperCase()),
-                avatar: {
-                  url: task?.assigneeProfilePicture || task?.creatorProfilePicture,
-                  isOwnerOfPod: false,
-                  color: null,
-                },
-              },
-            ]}
-            id={`task-${task?.id}`}
-          />
-          {task?.rewards?.length > 0 && <Compensation rewards={task?.rewards} />}
-        </TaskHeader>
-        <TaskContent>
-          <TaskTitle>{task?.title}</TaskTitle>
-          <TaskCardDescriptionText>
-            {renderMentionString({
-              content: task?.description,
-              router,
-            })}
-          </TaskCardDescriptionText>
-          {task?.podName && (
-            <PodWrapper
-              onClick={(e) => {
-                e.preventDefault();
-                e.stopPropagation();
-                const newUrl = `/pod/${task?.podId}/boards`;
-                window.location.href = newUrl;
-              }}
-            >
-              <PodName>{task?.podName.slice(0, 15)}</PodName>
-            </PodWrapper>
-          )}
-          {task?.media?.length > 0 ? <TaskMedia media={task?.media[0]} /> : <TaskSeparator />}
-        </TaskContent>
-        <TaskFooter>
-          {task?.changeRequestedAt && (
-            <div style={flexDivStyle}>
-              <RejectIcon style={rejectIconStyle} />
-              <TaskStatusHeaderText>Change requested</TaskStatusHeaderText>
-            </div>
-          )}
-          {task?.approvedAt && (
-            <div style={flexDivStyle}>
-              <CompletedIcon style={rejectIconStyle} />
-              <TaskStatusHeaderText>Approved</TaskStatusHeaderText>
-            </div>
-          )}
-          <TaskSummaryAction>
-            Details
-            <Arrow
-              style={{
-                marginLeft: '4px',
-              }}
-            />
-          </TaskSummaryAction>
-        </TaskFooter>
-      </TaskListCardWrapper>
-    </>
-  );
-}
+export default memo(Task);
