+++ wondrous-app/components/Common/Task/index.ts
@@ -1,0 +1,5 @@
+import Task from './Task';
+import type { TaskProps } from './Task';
+
+export default Task;
+export type { TaskProps };
