+++ wondrous-app/components/Common/Task/styles.tsx
@@ -1,11 +1,14 @@
-import { MenuItem, Modal, Typography } from '@mui/material';
-import ProposalIcon from 'components/Icons/proposalIcon';
 import styled from 'styled-components';
+import MenuItem from '@mui/material/MenuItem';
+import Modal from '@mui/material/Modal';
+import Typography from '@mui/material/Typography';
+
 import palette from 'theme/palette';
 
-import { CreateFormPreviewButton } from '../../CreateEntity/styles';
-import { BaseCard } from '../card';
-import { GradientMidnightDiagonal, GradientMidnightVertical } from '../gradients';
+import ProposalIcon from 'components/Icons/proposalIcon';
+import { CreateFormPreviewButton } from 'components/CreateEntity/styles';
+import { BaseCard } from 'components/Common/card';
+import { GradientMidnightDiagonal, GradientMidnightVertical } from 'components/Common/gradients';
 
 export const TaskInner = styled.div`
   display: flex;
@@ -403,12 +406,15 @@ export const ActionButton = styled(CreateFormPreviewButton)`
   }
 `;
 
-export const ProposalCardWrapper = styled.div`
+export const CardWrapper = styled.div`
+  padding-top: 1em;
+`;
+
+export const CardContent = styled.div`
   display: flex;
   flex-direction: column;
   color: ${({ theme }) => theme.palette.white};
   padding: 14px;
-  margin: ${(props) => (props.wrapped ? '0' : '1em 0 0 0')};
   border-radius: 6px;
   background: ${palette.grey900};
   gap: 14px;
@@ -522,7 +528,7 @@ export const TaskActionMenu = styled.div`
   z-index: 100;
   align-items: center;
   display: none;
-  ${ProposalCardWrapper}:hover & {
+  ${CardWrapper}:hover & {
     display: block;
   }
 `;
